[DFC-711][OLH-2469]: Remove UA code
Remove deprecated universal analytics code.
Create separate script for international phone number. Only include this
script on templates which have a phone number input field.

This is necessary as application.js has been removed
