+++ .env.sample
@@ -40,10 +40,8 @@ VERIFY_ACCESS_VALUE=
 # subjectId returned from oidc stub
 MY_ONE_LOGIN_USER_ID=user_id
 METHOD_MANAGEMENT_BASE_URL=https://method-management-v1-stub.home.dev.account.gov.uk
-UNIVERSAL_ANALYTICS_GTM_CONTAINER_ID="GTM-TK92W68"
 GOOGLE_ANALYTICS_4_GTM_CONTAINER_ID="GTM-KD86CMZ"
 GA4_ENABLED="true"
-UA_ENABLED="true"
 SELECT_TRACKING_ENABLED="true"
 SUPPORT_ADD_BACKUP_MFA="1"
 SUPPORT_METHOD_MANAGEMENT=1
