+++ package.json
@@ -7,14 +7,13 @@
   "scripts": {
     "bubbleprof": "NODE_ENV=production clinic bubbleprof --autocannon [ -c 10 -d 10 /healthcheck ] -- node -r dotenv/config dist/server.js && echo '\n\nRemember to clean up files in .clinic and node_trace.*.log\n'",
     "build-ts": "tsc",
-    "build": "npm run build-sass && npm run build-ts && npm run copy-assets && npm run build-js:analytics && npm run minfiy-build-js",
+    "build": "npm run build-sass && npm run build-ts && npm run copy-assets && npm run build-js:analytics",
     "build-dev": "npm run build-sass && npm run build-ts && npm run copy-assets",
     "build-sass": "rm -rf dist/public/style.css && sass --load-path=node_modules/govuk-frontend/dist/govuk --no-source-map src/assets/scss/application.scss dist/public/style.css --style compressed",
     "build-js:analytics": "cp node_modules/@govuk-one-login/frontend-analytics/lib/analytics.js dist/public/scripts",
     "clean": "rm -rf dist node_modules logs.json",
     "clean-modules": "rm -rf node_modules",
-    "concat-client-js": "mkdir -p dist/public/scripts && uglifyjs src/assets/javascript/cookies.js src/assets/javascript/application.js -o dist/public/scripts/application.js -b",
-    "copy-assets": "mkdir -p dist && npm run concat-client-js && copyfiles -u 1 src/**/*.njk dist/ src/locales/**/** dist/ src/config/*.txt dist/ && cp node_modules/govuk-frontend/dist/govuk/govuk-frontend.min.js dist/public/scripts",
+    "copy-assets": "mkdir -p dist/public/scripts && copyfiles -u 1 src/**/*.njk dist/ src/locales/**/** dist/ src/config/*.txt dist/ && cp node_modules/govuk-frontend/dist/govuk/govuk-frontend.min.js dist/public/scripts && cp src/assets/javascript/international-phone.js dist/public/scripts",
     "doctor": "NODE_ENV=production clinic doctor --autocannon [ -c 10 -d 10 /healthcheck ] -- node -r dotenv/config dist/server.js && echo '\n\nRemember to clean up files in .clinic and node_trace.*.log\n'",
     "deep-clean": "npm cache clear --force && npm cache clean --force && npm run clean && npm install && npm run build",
     "dev": "concurrently -k -p \"[{name}]\" -n \"Sass,TypeScript,Node\" -c \"yellow.bold,cyan.bold,green.bold\" \"npm run watch-sass\" \"npm run watch-ts\" \"npm run watch-node\"",
@@ -23,7 +22,6 @@
     "flame": "NODE_ENV=production clinic flame --autocannon [ -c 10 -d  10 /healthcheck ] -- node -r dotenv/config dist/server.js && echo '\n\nRemember to clean up files in .clinic and node_trace.*.log\n'",
     "heap-profiler": "NODE_ENV=production clinic flame --autocannon [ / --method GET -d 120 ] -- node -r dotenv/config dist/server.js && echo '\n\nRemember to clean up files in .clinic and node_trace.*.log\n'",
     "lint": "eslint .",
-    "minfiy-build-js": "uglifyjs dist/public/scripts/application.js -o dist/public/scripts/application.js -c -m",
     "start": "node dist/server.js",
     "test:audit": "npm audit",
     "test:integration": "rm -rf test/coverage && NODE_ENV=test nyc mocha -r dotenv/config -P tsconfig.mocha.json  \"src/**/*-integration.test.ts\"",
