+++ src/components/common/layout/base.njk
@@ -131,8 +131,6 @@
       initAll()
     </script>
 
-    <script type="text/javascript" src="/public/scripts/application.js" nonce='{{scriptNonce}}'></script>
-    <script type="text/javascript" src="/public/scripts/analytics.js" nonce='{{scriptNonce}}'></script>
     <script type="text/javascript" nonce='{{scriptNonce}}'>
 
       {% if backLinkJS %}
@@ -152,11 +150,9 @@
     <script type="text/javascript" {% if scriptNonce %} nonce="{{ scriptNonce }}"{%  endif %}>
     if (window.DI) {
       window.DI.appInit({
-        ga4ContainerId: "{{ga4ContainerId}}",
-        uaContainerId: "{{uaContainerId}}"
+        ga4ContainerId: "{{ga4ContainerId}}"
       }, {
         enableGa4Tracking: {{isGa4Enabled}},
-        enableUaTracking: {{isUaEnabled}},
         cookieDomain: "{{analyticsCookieDomain}}",
         isDataSensitive: true,
         enablePageViewTracking: true,
