+++ src/components/common/mfa/add-phone-number.njk
@@ -78,3 +78,6 @@
   </a>
   {{ ga4OnPageLoad({ nonce: scriptNonce,statusCode:"200",englishPageTitle:pageTitleName, taxonomyLevel1: "accounts", taxonomyLevel2:"change phone number", contentId:contentId,loggedInStatus:true,dynamic:true })}}
 {% endblock %}
+{% block scripts %}
+  <script type="text/javascript" src="/public/scripts/international-phone.js" nonce='{{scriptNonce}}'></script>
+{% endblock %}
