+++ src/components/common/sms/add-sms.njk
@@ -64,4 +64,8 @@
       "preventDoubleClick": true
     }) }}
     </form>
-{% endblock content %}
+{% endblock %}
+
+{% block scripts %}
+  <script type="text/javascript" src="/public/scripts/international-phone.js" nonce='{{scriptNonce}}'></script>
+{% endblock %}
