+++ src/utils/logger.ts
@@ -46,7 +46,6 @@ const ignorePaths = [
   "/public/scripts/all.js",
   "/public/style.css",
   "/public/style.css.map",
-  "/public/scripts/application.js",
   "/assets/images/govuk-crest-2x.png",
   "/assets/fonts/bold-b542beb274-v2.woff2",
   "/assets/fonts/bold-b542beb274-v2.woff2",
